fix: address linter errors
fix: configure notary once in node test

removes duplicate  call introduced by rebase

chore(ccq): update ccq metrics

chore(querystaking): update querystaking metrics

chore(query): update context handling in memorystore

fix(js-query): pass uint8Array buffer to BinaryReader

fix(query): use common.ReadSafe

+++ node/pkg/query/narasu/store.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Store interface {
-	IncrKey(ctx context.Context, bucket string, amount int, cur time.Time) (int, error)
-	GetKeys(ctx context.Context, bucket string, from time.Time, to time.Time) ([]int, error)
+	IncrKey(ctx context.Context, bucket string, amount int, cur time.Time) (uint64, error)
+	GetKeys(ctx context.Context, bucket string, from time.Time, to time.Time) ([]uint64, error)
 	Cleanup(ctx context.Context, now time.Time, age time.Duration) error
 }
